requirements.txt
streamlit>=1.28.0
neo4j>=5.14.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
groq>=0.4.0
asyncio>=3.4.3 